move python stuff to a src/ directory
