+++ domain-cc/svc-cc-j/src/rabbitmq_client.py
@@ -1,5 +1,5 @@
 import httpx
-from hoppy import Service as HoppyService
+from src.util.hoppy import Service as HoppyService
 
 RABBIT_MQ_CONFIG = {  # define this as a custom type
     "host": "localhost",
