+++ domain-cc/svc-cc-j/src/util/hoppy.py
